backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
Pillow==9.3.0
PyJWT==2.1.0
requests==2.26.0
django-filter==23.1
PyYAML==6.0
python-dotenv==0.19.0
gunicorn==20.1.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3